managing response based on requested id

+++ cashcard/src/main/java/com/example/cashcard/CashCard.java
@@ -1,3 +1,3 @@
 package com.example.cashcard;
 
-public record CashCard(Integer id, Double amount){}
+public record CashCard(Long id, Double amount){}
